Add a `README.markdown` file, pretty-formatted.
This commit adds a `README.markdown` file and fills it, primarily, with
the explanation of how to use the code given in the code comments. This
way, users will have a bit of an instructional guide for how to actually
make use of the program itself.

The `README.markdown` file in the root directory of a repository is a
convention that GitHub understands as the human-readable "introduction"
to the code contained within the repository. The `.markdown` extension
is the file format, sometimes abbreviated as `.md`. I like the longer,
more verbose file extension because it is less ambiguous.

This README also includes installation and running steps, which I simply
wrote down myself as I was figuring out how to run the code usefully.

There are no behavioral changes to the code in this commit. The code
works for me as-is, so this commit *only* adds a README file, it does
not change the runtime behavior at all.