Keep the "installed app" client ID and secret actually secret.
This commit removes the `participants.json` file, because it's being
used to distribute a specific OAuth 2.0 client ID and `client_secret`
value as part of the application. This is fine if you want everyone who
`git clone`s your repository to use *your* resource quota when hitting
the Google API, but is not fine if you want other folks who are cloning
your repository to have to use *their own* resource quota.

To address this, this commit replaces the `participants.json` file with
a `client_secrets.json` file and updates the `.gitignore` file to ensure
`client_secrets.json` is never committed to the repository from now on.

In order to actually use the code, however, other people will need to
create their own "Project" from the Google Developer API Console and
obtain their own client ID and secret file.